Fix getTypeOf and cloneDeep for instance of DataEntity
